[ADD] auto_backup: allow taking backup without having db manager publicly exposed
Before this commit there was only one solution to enable the auto_backup module: by having the database manager from Odoo exposed.

As we all know that's not the best idea, as documented at https://www.odoo.com/documentation/13.0/setup/deploy.html#db-manager-security

This commit introduces some very similar functions to take back-ups as the Odoo database manager does but without having to have the database manager exposed. Our functions will check if the user calling the function is the same user as authenticated in the cron and if the model name matches with the one we expect. This should disallow any malicious practices and allows to disable the database manager publicly.