Close sftp connection when an exception happens
When an exception happens, this addon leaves SFTP connection openned. And the next time that a backup is executed, it won't be able to put it into the SFTP server because there is an older SFTP connection openned.
So we added in  db_backup.py, a finally block  to the general try - except  that verifies if sftp object exists, then close sftp connection. 
Kind regards